+++ scripts/experiments.py
@@ -9,7 +9,6 @@ process_idx = args.idx
 # Automated grid search experiments
 seeds = [42, 43, 44, 45, 46]
 limits = [10, 30, 60, 100, 150, 250, 500, 700, 1000, 1300, 1700, 2000, 2500, 3000, 0]
-# limits = [80, 100, 120, 1000, 1500, 2300, 2900]
 bert_names = [
     "/export/home/pwajsburt/data/models/embedding-whole-word/checkpoint-250000/",
     "/export/home/share/datascientists/models/camembert-base",
@@ -17,10 +16,10 @@ bert_names = [
 ]
 doc_ablations = [
     "'doc._.note_class_source_value != \"CR-ACTE-DIAG-AUTRE\"'",
-    # "'doc._.note_class_source_value != \"CR-ANAPATH\"'",
-    # "'doc._.note_class_source_value != \"CR-IMAGE\"'",
-    # "'doc._.note_class_source_value != \"CR-OPER\"'",
-    # "'doc._.note_class_source_value != \"RCP\"'",
+    "'doc._.note_class_source_value != \"CR-ANAPATH\"'",
+    "'doc._.note_class_source_value != \"CR-IMAGE\"'",
+    "'doc._.note_class_source_value != \"CR-OPER\"'",
+    "'doc._.note_class_source_value != \"RCP\"'",
 ]
 
 # Iterate over all combinations of hyperparameter values.
@@ -44,40 +43,40 @@ for seed in seeds:
             subprocess.run(xp)
         i += 1
 
-#    for bert_name in bert_names:
-#        xp = [
-#            "dvc",
-#            "exp",
-#            "run",
-#            "--queue",
-#            #"--temp",
-#            "-S",
-#            f"configs/config.cfg:system.seed={seed}",
-#            "-S",
-#            f"configs/config.cfg:paths.bert={bert_name}",
-#        ]
-#        if process_idx is None or i % process_idx == 0:
-#            print("Running", " ".join(xp))
-#            subprocess.run(xp)
-#        i += 1
-#
-#    ## Iterate over all combinations of hyperparameter values.
-#    for limit in limits:
-#        if limit == 0:
-#            # We already performed the xp with no limit above
-#            continue
-#        xp = [
-#            "dvc",
-#            "exp",
-#            "run",
-#            "--queue",
-#            #"--temp",
-#            "-S",
-#            f"configs/config.cfg:system.seed={seed}",
-#            "-S",
-#            f"configs/config.cfg:corpora.train.limit={limit}",
-#        ]
-#        if process_idx is None or i % process_idx == 0:
-#            print("Running", " ".join(xp))
-#            subprocess.run(xp)
-#        i += 1
+    for bert_name in bert_names:
+        xp = [
+            "dvc",
+            "exp",
+            "run",
+            "--queue",
+            # "--temp",
+            "-S",
+            f"configs/config.cfg:system.seed={seed}",
+            "-S",
+            f"configs/config.cfg:paths.bert={bert_name}",
+        ]
+        if process_idx is None or i % process_idx == 0:
+            print("Running", " ".join(xp))
+            subprocess.run(xp)
+        i += 1
+
+    ## Iterate over all combinations of hyperparameter values.
+    for limit in limits:
+        if limit == 0:
+            # We already performed the xp with no limit above
+            continue
+        xp = [
+            "dvc",
+            "exp",
+            "run",
+            "--queue",
+            # "--temp",
+            "-S",
+            f"configs/config.cfg:system.seed={seed}",
+            "-S",
+            f"configs/config.cfg:corpora.train.limit={limit}",
+        ]
+        if process_idx is None or i % process_idx == 0:
+            print("Running", " ".join(xp))
+            subprocess.run(xp)
+        i += 1